Moved API call to APP.jsx
